feat: support v7 raw stream chunks and raw usage
Two AI SDK v7 provider-spec features the Ollama provider did not implement:

- `includeRawChunks`: doStream emits `{ type: 'raw', rawValue }` parts carrying
  each unmodified Ollama chunk. Off by default.
- `usage.raw`: doGenerate/doStream expose Ollama's own counters and nanosecond
  timings alongside the normalised token counts.

The streaming `finish` part now carries the same `providerMetadata` doGenerate
already returned.

Decomposition
-------------

The two files past 1000 lines are split along their real seams, no behaviour
change. chat-language-model.ts (1645 -> 714) keeps the LanguageModelV4
implementation and reliability orchestration:

- chat-request.ts   - call options, think resolution, schema cleaning
- chat-result.ts    - content/usage/metadata assembly for both result paths
- chat-stream.ts    - the Ollama chunk -> stream part state machine

object-generation-reliability.ts (1711 -> 461) keeps recovery orchestration:

- json-text-repair.ts      - the repair passes and parse-with-repair
- json-schema-coercion.ts  - schema fallbacks and type coercion

enhancedRepairText, cascadeRepairText and RepairTextFunction are still exported
from the package root; only internal module paths changed. The chat model test
file split along the same seam, with its fixtures moved to
chat-language-model.test-helpers.ts rather than duplicated.

Duplication and dead code removed
---------------------------------

- The providerMetadata.ollama payload was hand-rolled in four places, three with
  the same `if (x !== undefined)` ladder and a cast to paper over undefined. One
  ollamaResponseDetails helper now; the cast is gone because it omits absent
  keys instead.
- doGenerate hand-rolled the same content/metadata/request payload that
  buildGenerationResult already assembled; it now calls it. buildGenerationResult
  derives the tool-calls finish reason from parsedToolCalls rather than taking it
  on faith from the caller.
- The stream transform emitted text, reasoning and tool calls twice, once per
  branch of `chunk.done`. Both branches now drive shared emitters.
- Two dead `?? 'stop'` fallbacks behind a cast: mapOllamaFinishReason already
  defaults a missing reason to stop, and a bare 'stop' string was never a valid
  V4 finish reason.
- getRepairFunction took the whole ObjectGenerationOptions to read two fields;
  it now takes those two fields, which also breaks the import cycle the split
  would otherwise have created.
- buildObjectGenerationResult declared a validatedObject parameter it never read.

Dependency fixes
----------------

Brings the bump on this branch to a state that installs and builds:

- Caret floors set to the newest versions older than the 7-day
  minimumReleaseAge window, which otherwise fails resolution.
- typescript held at ^6.0.3: TS 7 breaks tsup 8.5.1's dts build via
  rollup-plugin-dts.
- reranking-model tests use real `Response` objects; provider-utils >= 5.0.2
  reads `response.body`, so the old `{ text() }` stubs decoded to "".
- Two new unicorn 72 condition-ordering rules disabled; reordering `&&`
  operands changes short-circuit behaviour.
