+++ examples/browser/package.json
@@ -16,25 +16,25 @@
     "quality": "pnpm type-check && pnpm build"
   },
   "dependencies": {
-    "@ai-sdk/react": "^4.0.8",
-    "@radix-ui/react-avatar": "^1.2.0",
-    "@radix-ui/react-collapsible": "^1.1.14",
-    "@radix-ui/react-dropdown-menu": "^2.1.18",
-    "@radix-ui/react-hover-card": "^1.1.17",
-    "@radix-ui/react-progress": "^1.1.10",
-    "@radix-ui/react-scroll-area": "^1.2.12",
-    "@radix-ui/react-select": "^2.3.1",
+    "@ai-sdk/react": "^4.0.34",
+    "@radix-ui/react-avatar": "^1.2.2",
+    "@radix-ui/react-collapsible": "^1.1.16",
+    "@radix-ui/react-dropdown-menu": "^2.1.20",
+    "@radix-ui/react-hover-card": "^1.1.19",
+    "@radix-ui/react-progress": "^1.1.12",
+    "@radix-ui/react-scroll-area": "^1.2.14",
+    "@radix-ui/react-select": "^2.3.3",
     "@radix-ui/react-slot": "^1.3.0",
-    "@radix-ui/react-tooltip": "^1.2.10",
+    "@radix-ui/react-tooltip": "^1.2.12",
     "@radix-ui/react-use-controllable-state": "^1.2.3",
-    "@tailwindcss/vite": "^4.3.2",
-    "ai": "^7.0.4",
+    "@tailwindcss/vite": "^4.3.3",
+    "ai": "^7.0.31",
     "ai-sdk-ollama": "workspace:*",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "embla-carousel-react": "^8.6.0",
-    "lucide-react": "^1.22.0",
-    "nanoid": "^5.1.16",
+    "lucide-react": "^1.25.0",
+    "nanoid": "^6.0.0",
     "react": "^19.2.7",
     "react-dom": "^19.2.7",
     "react-syntax-highlighter": "^16.1.1",
@@ -46,13 +46,13 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.61.1",
-    "@types/node": "^26.0.1",
+    "@types/node": "^26.1.1",
     "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
     "@types/react-syntax-highlighter": "^15.5.13",
     "rimraf": "^6.1.3",
-    "tailwindcss": "^4.3.2",
+    "tailwindcss": "^4.3.3",
     "typescript": "^6.0.3",
-    "vite": "^8.1.0"
+    "vite": "^8.1.5"
   }
 }
