+++ examples/node/package.json
@@ -8,17 +8,17 @@
     "quality": "pnpm build"
   },
   "dependencies": {
-    "@ai-sdk/mcp": "^2.0.0",
-    "ai": "^7.0.4",
+    "@ai-sdk/mcp": "^2.0.15",
+    "ai": "^7.0.31",
     "ai-sdk-ollama": "workspace:*",
     "dotenv": "^17.4.2",
     "ollama": "^0.6.3",
     "zod": "^4.4.3"
   },
   "devDependencies": {
-    "@types/node": "^26.0.1",
+    "@types/node": "^26.1.1",
     "rimraf": "^6.1.3",
-    "tsx": "^4.22.4",
+    "tsx": "^4.23.1",
     "typescript": "^6.0.3"
   }
 }
