+++ package.json
@@ -22,16 +22,16 @@
     "quality": "turbo quality"
   },
   "devDependencies": {
-    "@changesets/cli": "^2.31.0",
+    "@changesets/cli": "^2.31.1",
     "@eslint/js": "^10.0.1",
-    "@typescript-eslint/eslint-plugin": "^8.61.1",
-    "@typescript-eslint/parser": "^8.62.1",
-    "eslint": "^10.6.0",
+    "@typescript-eslint/eslint-plugin": "^8.64.0",
+    "@typescript-eslint/parser": "^8.64.0",
+    "eslint": "^10.7.0",
     "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-unicorn": "^69.0.0",
-    "prettier": "^3.9.1",
-    "turbo": "^2.10.1",
-    "typescript-eslint": "^8.62.1"
+    "eslint-plugin-unicorn": "^72.0.0",
+    "prettier": "^3.9.5",
+    "turbo": "^2.10.5",
+    "typescript-eslint": "^8.64.0"
   },
   "engines": {
     "node": ">=22",
