+++ packages/ai-sdk-ollama/package.json
@@ -84,32 +84,32 @@
     "tool-calling"
   ],
   "dependencies": {
-    "@ai-sdk/provider": "^4.0.0",
-    "@ai-sdk/provider-utils": "^5.0.1",
-    "jsonrepair": "^3.14.0",
+    "@ai-sdk/provider": "^4.0.3",
+    "@ai-sdk/provider-utils": "^5.0.11",
+    "jsonrepair": "^3.15.0",
     "ollama": "^0.6.3"
   },
   "peerDependencies": {
-    "ai": "^7.0.0"
+    "ai": "^7.0.31"
   },
   "devDependencies": {
     "@ai-sdk/test-server": "^2.0.0",
-    "@arethetypeswrong/cli": "^0.18.4",
+    "@arethetypeswrong/cli": "^0.18.5",
     "@edge-runtime/vm": "^5.0.0",
     "@total-typescript/ts-reset": "^0.6.1",
     "@total-typescript/tsconfig": "^1.0.4",
     "@types/eslint-config-prettier": "^6.11.3",
-    "@types/node": "^26.0.1",
-    "@typescript-eslint/eslint-plugin": "^8.61.1",
-    "@typescript-eslint/parser": "^8.62.1",
+    "@types/node": "^26.1.1",
+    "@typescript-eslint/eslint-plugin": "^8.64.0",
+    "@typescript-eslint/parser": "^8.64.0",
     "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-unicorn": "^69.0.0",
-    "prettier": "^3.9.1",
+    "eslint-plugin-unicorn": "^72.0.0",
+    "prettier": "^3.9.5",
     "rimraf": "^6.1.3",
     "tsup": "^8.5.1",
     "typescript": "^6.0.3",
-    "typescript-eslint": "^8.62.1",
-    "vitest": "^4.1.8",
+    "typescript-eslint": "^8.64.0",
+    "vitest": "^4.1.10",
     "zod": "^4.4.3"
   },
   "engines": {
