+++ packages/ai-sdk-ollama/src/models/chat-language-model.test.ts
@@ -1,55 +1,23 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { OllamaChatLanguageModel } from './chat-language-model';
-import { OllamaChatSettings } from '../provider';
 import {
   LanguageModelV4CallOptions,
-  LanguageModelV4StreamPart,
   LanguageModelV4FunctionTool,
 } from '@ai-sdk/provider';
-import { Ollama, AbortableAsyncIterator, ChatResponse } from 'ollama';
-import { convertArrayToAsyncIterable } from '@ai-sdk/provider-utils/test';
+import { ChatResponse } from 'ollama';
+import { OllamaChatSettings } from '../provider';
+import {
+  createExpectedUsage,
+  createModel,
+  mockOllamaClient,
+} from './chat-language-model.test-helpers';
 
-// Mock Ollama client
-const mockOllamaClient = {
-  chat: vi.fn(),
-} as unknown as Ollama;
-
-// Helper to mock streaming chat responses (consistent with AI SDK provider patterns)
-function mockChatStream(data: ChatResponse[]): void {
-  const stream = convertArrayToAsyncIterable(data);
-  (mockOllamaClient.chat as ReturnType<typeof vi.fn>).mockResolvedValueOnce(
-    stream as unknown as AbortableAsyncIterator<ChatResponse>,
-  );
-}
-
-// Helper to create V3 usage format for test expectations
-function createExpectedUsage(inputTokens: number, outputTokens: number) {
-  return {
-    inputTokens: {
-      total: inputTokens,
-      noCache: inputTokens,
-      cacheRead: undefined,
-      cacheWrite: undefined,
-    },
-    outputTokens: {
-      total: outputTokens,
-      text: outputTokens,
-      reasoning: undefined,
-    },
-  };
-}
-
-describe('OllamaChatLanguageModel', () => {
+describe('OllamaChatLanguageModel: doGenerate', () => {
   let model: OllamaChatLanguageModel;
-  let settings: OllamaChatSettings;
 
   beforeEach(() => {
     vi.clearAllMocks();
-    settings = {};
-    model = new OllamaChatLanguageModel('llama3.2', settings, {
-      client: mockOllamaClient,
-      provider: 'ollama',
-    });
+    model = createModel();
   });
 
   describe('initialization', () => {
@@ -117,6 +85,35 @@ describe('OllamaChatLanguageModel', () => {
         messages: [{ role: 'user', content: 'Hello' }],
         stream: false,
         options: expect.any(Object),
+      });
+    });
+
+    it('should expose Ollama token counts and timings on usage.raw', async () => {
+      vi.mocked(mockOllamaClient.chat).mockResolvedValueOnce({
+        model: 'llama3.2',
+        created_at: new Date(),
+        message: { role: 'assistant', content: 'Hello, world!' },
+        done: true,
+        done_reason: 'stop',
+        eval_count: 10,
+        prompt_eval_count: 5,
+        total_duration: 1_000_000_000,
+        load_duration: 100_000_000,
+        prompt_eval_duration: 200_000_000,
+        eval_duration: 700_000_000,
+      } as unknown as ChatResponse);
+
+      const result = await model.doGenerate({
+        prompt: [{ role: 'user', content: [{ type: 'text', text: 'Hello' }] }],
+      });
+
+      expect(result.usage.raw).toEqual({
+        prompt_eval_count: 5,
+        eval_count: 10,
+        total_duration: 1_000_000_000,
+        load_duration: 100_000_000,
+        prompt_eval_duration: 200_000_000,
+        eval_duration: 700_000_000,
       });
     });
 
@@ -481,237 +478,6 @@ describe('OllamaChatLanguageModel', () => {
     });
   });
 
-  describe('doStream', () => {
-    it('should handle streaming responses', async () => {
-      const mockStreamData = [
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: 'Hello' },
-          done: false,
-          done_reason: '',
-          eval_count: 5,
-          prompt_eval_count: 3,
-          total_duration: 500_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 350_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: ' world' },
-          done: false,
-          done_reason: '',
-          eval_count: 10,
-          prompt_eval_count: 3,
-          total_duration: 800_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 650_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: '!' },
-          done: true,
-          done_reason: 'stop',
-          eval_count: 15,
-          prompt_eval_count: 8,
-          total_duration: 1_000_000_000,
-          load_duration: 100_000_000,
-          prompt_eval_duration: 200_000_000,
-          eval_duration: 700_000_000,
-        },
-      ];
-
-      mockChatStream(mockStreamData);
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [{ role: 'user', content: [{ type: 'text', text: 'Hello' }] }],
-      };
-
-      const { stream } = await model.doStream(options);
-      const chunks = await Array.fromAsync(stream);
-
-      expect(chunks).toHaveLength(7); // stream-start + text-start + 3 text-delta + text-end + finish
-      // V3 adds stream-start at the beginning
-      expect(chunks[0]).toEqual({
-        type: 'stream-start',
-        warnings: [],
-      });
-      expect(chunks[1]).toEqual({
-        type: 'text-start',
-        id: expect.any(String),
-      });
-      const textStartId = (chunks[1] as { id: string }).id;
-      expect(chunks[2]).toEqual({
-        type: 'text-delta',
-        id: textStartId,
-        delta: 'Hello',
-      });
-      expect(chunks[3]).toEqual({
-        type: 'text-delta',
-        id: textStartId,
-        delta: ' world',
-      });
-      expect(chunks[4]).toEqual({
-        type: 'text-delta',
-        id: textStartId,
-        delta: '!',
-      });
-      expect(chunks[5]).toEqual({
-        type: 'text-end',
-        id: textStartId,
-      });
-      expect(chunks[6]).toEqual({
-        type: 'finish',
-        finishReason: { unified: 'stop', raw: 'stop' },
-        usage: createExpectedUsage(8, 15),
-      });
-
-      expect(mockOllamaClient.chat).toHaveBeenCalledWith({
-        model: 'llama3.2',
-        messages: [{ role: 'user', content: 'Hello' }],
-        stream: true,
-        options: expect.objectContaining({}),
-      });
-    });
-
-    it('should return tool-calls finish reason when stream emits tool calls', async () => {
-      const mockStreamData = [
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: {
-            role: 'assistant',
-            content: '',
-            tool_calls: [
-              {
-                function: {
-                  name: 'get_weather',
-                  arguments: { location: 'Paris' },
-                },
-              },
-            ],
-          },
-          done: false,
-          done_reason: '',
-          eval_count: 5,
-          prompt_eval_count: 3,
-          total_duration: 500_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 350_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: '' },
-          done: true,
-          done_reason: 'stop',
-          eval_count: 10,
-          prompt_eval_count: 5,
-          total_duration: 1_000_000_000,
-          load_duration: 100_000_000,
-          prompt_eval_duration: 200_000_000,
-          eval_duration: 700_000_000,
-        },
-      ];
-
-      mockChatStream(mockStreamData);
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [
-          {
-            role: 'user',
-            content: [{ type: 'text', text: 'What is the weather?' }],
-          },
-        ],
-        tools: [
-          {
-            type: 'function',
-            name: 'get_weather',
-            description: 'Get weather for a city',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                location: { type: 'string' },
-              },
-            },
-          },
-        ],
-      };
-
-      const { stream } = await model.doStream(options);
-      const chunks = await Array.fromAsync(stream);
-
-      expect(chunks).toEqual([
-        {
-          type: 'stream-start',
-          warnings: [],
-        },
-        {
-          type: 'tool-call',
-          toolCallId: expect.any(String),
-          toolName: 'get_weather',
-          input: JSON.stringify({ location: 'Paris' }),
-        },
-        {
-          type: 'finish',
-          finishReason: {
-            unified: 'tool-calls',
-            raw: 'stop',
-          },
-          usage: createExpectedUsage(5, 10),
-        },
-      ]);
-    });
-
-    it('should handle streaming errors', async () => {
-      const error = new Error('Stream error');
-      vi.mocked(mockOllamaClient.chat).mockRejectedValueOnce(error);
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [{ role: 'user', content: [{ type: 'text', text: 'Hello' }] }],
-      };
-
-      await expect(model.doStream(options)).rejects.toThrow('Stream error');
-    });
-
-    it('should handle abort signal in streaming', async () => {
-      const abortController = new AbortController();
-      const mockAsyncIterable = {
-        [Symbol.asyncIterator]: vi.fn().mockReturnValue({
-          next: vi.fn().mockImplementation(async () => {
-            abortController.abort();
-            throw new Error('Aborted');
-          }),
-        }),
-      };
-
-      (
-        mockOllamaClient.chat as unknown as ReturnType<typeof vi.fn>
-      ).mockResolvedValueOnce(
-        mockAsyncIterable as unknown as AbortableAsyncIterator<ChatResponse>,
-      );
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [{ role: 'user', content: [{ type: 'text', text: 'Hello' }] }],
-        abortSignal: abortController.signal,
-      };
-
-      const { stream } = await model.doStream(options);
-
-      await expect(async () => {
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        for await (const _chunk of stream) {
-          // Should throw before getting here
-        }
-      }).rejects.toThrow('Aborted');
-    });
-  });
-
   describe('with custom settings', () => {
     it('should use custom Ollama options', async () => {
       const customSettings: OllamaChatSettings = {
@@ -917,600 +683,6 @@ describe('OllamaChatLanguageModel', () => {
         { type: 'text', text: 'The answer is 42.' },
       ]);
       expect(result.finishReason).toEqual({ unified: 'stop', raw: 'stop' });
-    });
-  });
-
-  describe('doStream with think', () => {
-    it('should handle streaming with reasoning', async () => {
-      const mockStreamData: ChatResponse[] = [
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: {
-            role: 'assistant',
-            content: '',
-            thinking: 'Let me think about this step by step.',
-          },
-          done: false,
-          done_reason: 'stop',
-          eval_count: 5,
-          prompt_eval_count: 3,
-          total_duration: 500_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 350_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: {
-            role: 'assistant',
-            content: 'The answer is 42.',
-            thinking: '',
-          },
-          done: true,
-          done_reason: 'stop',
-          eval_count: 10,
-          prompt_eval_count: 3,
-          total_duration: 1_000_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 850_000_000,
-        },
-      ];
-
-      mockChatStream(mockStreamData);
-
-      const modelWithReasoning = new OllamaChatLanguageModel(
-        'llama3.2',
-        { think: true },
-        { client: mockOllamaClient, provider: 'ollama' },
-      );
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [
-          {
-            role: 'user',
-            content: [{ type: 'text', text: 'What is the answer?' }],
-          },
-        ],
-      };
-
-      const { stream } = await modelWithReasoning.doStream(options);
-      const chunks: LanguageModelV4StreamPart[] = await Array.fromAsync(stream);
-
-      // Check that reasoning stream parts are emitted
-      const reasoningStart = chunks.find(
-        (part) => part.type === 'reasoning-start',
-      );
-      const reasoningDelta = chunks.find(
-        (part) => part.type === 'reasoning-delta',
-      );
-      const reasoningEnd = chunks.find((part) => part.type === 'reasoning-end');
-      const textDelta = chunks.find((part) => part.type === 'text-delta');
-      const finish = chunks.find((part) => part.type === 'finish');
-
-      expect(reasoningStart).toBeDefined();
-      expect(reasoningDelta).toBeDefined();
-      expect(reasoningDelta?.delta).toBe(
-        'Let me think about this step by step.',
-      );
-      expect(reasoningEnd).toBeDefined();
-
-      // Verify that all reasoning events use the same ID
-      // This is critical for AI SDK to properly aggregate reasoning content
-      expect(reasoningStart?.id).toBeDefined();
-      expect(reasoningDelta?.id).toBeDefined();
-      expect(reasoningEnd?.id).toBeDefined();
-      expect(reasoningStart?.id).toBe(reasoningDelta?.id);
-      expect(reasoningDelta?.id).toBe(reasoningEnd?.id);
-      expect(reasoningStart?.id).toBe(reasoningEnd?.id);
-
-      // Final text may be emitted on the done chunk; accept either behavior
-      // If text was emitted, it should match our final chunk content
-      if (textDelta) {
-        expect(textDelta).toEqual({
-          type: 'text-delta',
-          id: expect.any(String),
-          delta: 'The answer is 42.',
-        });
-      }
-      expect(finish).toBeDefined();
-      expect(finish?.finishReason).toEqual({ unified: 'stop', raw: 'stop' });
-    });
-
-    it('should not emit reasoning stream parts when think is disabled', async () => {
-      const mockStreamData: ChatResponse[] = [
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: {
-            role: 'assistant',
-            content: 'The answer is 42.',
-            thinking: 'Let me think about this step by step.',
-          },
-          done: true,
-          done_reason: 'stop',
-          eval_count: 10,
-          prompt_eval_count: 3,
-          total_duration: 1_000_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 850_000_000,
-        },
-      ];
-
-      mockChatStream(mockStreamData);
-
-      const modelWithoutReasoning = new OllamaChatLanguageModel(
-        'llama3.2',
-        { think: false },
-        { client: mockOllamaClient, provider: 'ollama' },
-      );
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [
-          {
-            role: 'user',
-            content: [{ type: 'text', text: 'What is the answer?' }],
-          },
-        ],
-      };
-
-      const { stream } = await modelWithoutReasoning.doStream(options);
-      const chunks: LanguageModelV4StreamPart[] = await Array.fromAsync(stream);
-
-      // Check that reasoning stream parts are NOT emitted
-      const reasoningStart = chunks.find(
-        (part) => part.type === 'reasoning-start',
-      );
-      const reasoningDelta = chunks.find(
-        (part) => part.type === 'reasoning-delta',
-      );
-      const reasoningEnd = chunks.find((part) => part.type === 'reasoning-end');
-      const textDelta = chunks.find((part) => part.type === 'text-delta');
-      const finish = chunks.find((part) => part.type === 'finish');
-
-      expect(reasoningStart).toBeUndefined();
-      expect(reasoningDelta).toBeUndefined();
-      expect(reasoningEnd).toBeUndefined();
-      // Final text may be emitted on the done chunk; accept either behavior
-      if (textDelta) {
-        expect(textDelta).toEqual({
-          type: 'text-delta',
-          id: expect.any(String),
-          delta: 'The answer is 42.',
-        });
-      }
-      expect(finish).toBeDefined();
-      expect(finish?.finishReason).toEqual({ unified: 'stop', raw: 'stop' });
-    });
-
-    it('should emit a single reasoning-start/end cycle for multiple reasoning chunks', async () => {
-      // Simulate how models like Qwen 3/DeepSeek-R1 stream reasoning tokens
-      // chunk-by-chunk (each chunk has a thinking field with a small piece)
-      const mockStreamData: ChatResponse[] = [
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: {
-            role: 'assistant',
-            content: '',
-            thinking: 'Let me ',
-          },
-          done: false,
-          done_reason: 'stop',
-          eval_count: 2,
-          prompt_eval_count: 3,
-          total_duration: 200_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 100_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: {
-            role: 'assistant',
-            content: '',
-            thinking: 'think about ',
-          },
-          done: false,
-          done_reason: 'stop',
-          eval_count: 4,
-          prompt_eval_count: 3,
-          total_duration: 400_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 300_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: {
-            role: 'assistant',
-            content: '',
-            thinking: 'this.',
-          },
-          done: false,
-          done_reason: 'stop',
-          eval_count: 5,
-          prompt_eval_count: 3,
-          total_duration: 500_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 350_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: {
-            role: 'assistant',
-            content: 'The answer is 4.',
-            thinking: '',
-          },
-          done: true,
-          done_reason: 'stop',
-          eval_count: 10,
-          prompt_eval_count: 3,
-          total_duration: 1_000_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 850_000_000,
-        },
-      ];
-
-      mockChatStream(mockStreamData);
-
-      const modelWithReasoning = new OllamaChatLanguageModel(
-        'llama3.2',
-        { think: true },
-        { client: mockOllamaClient, provider: 'ollama' },
-      );
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [
-          {
-            role: 'user',
-            content: [{ type: 'text', text: 'What is 2 + 2?' }],
-          },
-        ],
-      };
-
-      const { stream } = await modelWithReasoning.doStream(options);
-      const chunks: LanguageModelV4StreamPart[] = await Array.fromAsync(stream);
-
-      // There should be exactly ONE reasoning-start and ONE reasoning-end
-      const reasoningStarts = chunks.filter(
-        (part) => part.type === 'reasoning-start',
-      );
-      const reasoningDeltas = chunks.filter(
-        (part) => part.type === 'reasoning-delta',
-      );
-      const reasoningEnds = chunks.filter(
-        (part) => part.type === 'reasoning-end',
-      );
-
-      expect(reasoningStarts).toHaveLength(1);
-      expect(reasoningEnds).toHaveLength(1);
-      // Three reasoning chunks should produce three deltas
-      expect(reasoningDeltas).toHaveLength(3);
-
-      // All reasoning events must share the same ID
-      const reasoningId = reasoningStarts[0]?.id;
-      expect(reasoningId).toBeDefined();
-      for (const delta of reasoningDeltas) {
-        expect(delta.id).toBe(reasoningId);
-      }
-      expect(reasoningEnds[0]?.id).toBe(reasoningId);
-
-      // Verify the delta content
-      expect(reasoningDeltas[0]?.delta).toBe('Let me ');
-      expect(reasoningDeltas[1]?.delta).toBe('think about ');
-      expect(reasoningDeltas[2]?.delta).toBe('this.');
-
-      // Verify text content follows reasoning
-      const textDelta = chunks.find((part) => part.type === 'text-delta');
-      if (textDelta) {
-        expect(textDelta).toEqual({
-          type: 'text-delta',
-          id: expect.any(String),
-          delta: 'The answer is 4.',
-        });
-      }
-
-      const finish = chunks.find((part) => part.type === 'finish');
-      expect(finish).toBeDefined();
-      expect(finish?.finishReason).toEqual({ unified: 'stop', raw: 'stop' });
-    });
-  });
-
-  describe('UI Message Stream compatibility', () => {
-    it('should emit text-start, text-delta, and text-end for UI message streaming', async () => {
-      const mockStreamData = [
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: 'Hello' },
-          done: false,
-          done_reason: '',
-          eval_count: 5,
-          prompt_eval_count: 3,
-          total_duration: 500_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 350_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: ' world' },
-          done: false,
-          done_reason: '',
-          eval_count: 10,
-          prompt_eval_count: 3,
-          total_duration: 800_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 650_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: '!' },
-          done: true,
-          done_reason: 'stop',
-          eval_count: 15,
-          prompt_eval_count: 8,
-          total_duration: 1_000_000_000,
-          load_duration: 100_000_000,
-          prompt_eval_duration: 200_000_000,
-          eval_duration: 700_000_000,
-        },
-      ];
-
-      mockChatStream(mockStreamData);
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [{ role: 'user', content: [{ type: 'text', text: 'Hello' }] }],
-      };
-
-      const { stream } = await model.doStream(options);
-      const chunks: LanguageModelV4StreamPart[] = await Array.fromAsync(stream);
-
-      // Should have: stream-start, text-start, text-delta, text-delta, text-delta, text-end, finish
-      expect(chunks).toHaveLength(7);
-
-      // V3 adds stream-start at the beginning
-      expect(chunks[0]).toEqual({
-        type: 'stream-start',
-        warnings: [],
-      });
-
-      // Check text-start is emitted second
-      expect(chunks[1]).toEqual({
-        type: 'text-start',
-        id: expect.any(String),
-      });
-
-      // Check text-delta parts have the same ID
-      const textStartId = (chunks[1] as { id: string }).id;
-      expect(chunks[2]).toEqual({
-        type: 'text-delta',
-        id: textStartId,
-        delta: 'Hello',
-      });
-      expect(chunks[3]).toEqual({
-        type: 'text-delta',
-        id: textStartId,
-        delta: ' world',
-      });
-      expect(chunks[4]).toEqual({
-        type: 'text-delta',
-        id: textStartId,
-        delta: '!',
-      });
-
-      // Check text-end is emitted with the same ID
-      expect(chunks[5]).toEqual({
-        type: 'text-end',
-        id: textStartId,
-      });
-
-      // Check finish is emitted last
-      expect(chunks[6]).toEqual({
-        type: 'finish',
-        finishReason: { unified: 'stop', raw: 'stop' },
-        usage: createExpectedUsage(8, 15),
-      });
-    });
-
-    it('should handle empty content gracefully without text-start/end', async () => {
-      const mockStreamData = [
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: '' },
-          done: true,
-          done_reason: 'stop',
-          eval_count: 0,
-          prompt_eval_count: 5,
-          total_duration: 1_000_000_000,
-          load_duration: 100_000_000,
-          prompt_eval_duration: 200_000_000,
-          eval_duration: 700_000_000,
-        },
-      ];
-
-      mockChatStream(mockStreamData);
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [{ role: 'user', content: [{ type: 'text', text: 'Test' }] }],
-      };
-
-      const { stream } = await model.doStream(options);
-      const chunks: LanguageModelV4StreamPart[] = await Array.fromAsync(stream);
-
-      // Should have stream-start + finish (no text parts since content is empty)
-      expect(chunks).toHaveLength(2);
-      expect(chunks[0]).toEqual({
-        type: 'stream-start',
-        warnings: [],
-      });
-      expect(chunks[1]).toEqual({
-        type: 'finish',
-        finishReason: { unified: 'stop', raw: 'stop' },
-        usage: createExpectedUsage(5, 0),
-      });
-    });
-
-    it('should handle content only in final chunk', async () => {
-      const mockStreamData = [
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: '' },
-          done: false,
-          done_reason: '',
-          eval_count: 5,
-          prompt_eval_count: 3,
-          total_duration: 500_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 350_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: 'Complete response' },
-          done: true,
-          done_reason: 'stop',
-          eval_count: 15,
-          prompt_eval_count: 8,
-          total_duration: 1_000_000_000,
-          load_duration: 100_000_000,
-          prompt_eval_duration: 200_000_000,
-          eval_duration: 700_000_000,
-        },
-      ];
-
-      mockChatStream(mockStreamData);
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [{ role: 'user', content: [{ type: 'text', text: 'Test' }] }],
-      };
-
-      const { stream } = await model.doStream(options);
-      const chunks: LanguageModelV4StreamPart[] = await Array.fromAsync(stream);
-
-      // Should have: stream-start, text-start, text-delta, text-end, finish
-      expect(chunks).toHaveLength(5);
-
-      // V3 adds stream-start at the beginning
-      expect(chunks[0]).toEqual({
-        type: 'stream-start',
-        warnings: [],
-      });
-
-      // Check text-start is emitted when content appears in final chunk
-      expect(chunks[1]).toEqual({
-        type: 'text-start',
-        id: expect.any(String),
-      });
-
-      const textStartId = (chunks[1] as { id: string }).id;
-      expect(chunks[2]).toEqual({
-        type: 'text-delta',
-        id: textStartId,
-        delta: 'Complete response',
-      });
-
-      expect(chunks[3]).toEqual({
-        type: 'text-end',
-        id: textStartId,
-      });
-
-      expect(chunks[4]).toEqual({
-        type: 'finish',
-        finishReason: { unified: 'stop', raw: 'stop' },
-        usage: createExpectedUsage(8, 15),
-      });
-    });
-
-    it('should maintain consistent text ID across all text stream parts', async () => {
-      const mockStreamData = [
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: 'Part' },
-          done: false,
-          done_reason: '',
-          eval_count: 5,
-          prompt_eval_count: 3,
-          total_duration: 500_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 350_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: ' two' },
-          done: false,
-          done_reason: '',
-          eval_count: 10,
-          prompt_eval_count: 3,
-          total_duration: 800_000_000,
-          load_duration: 50_000_000,
-          prompt_eval_duration: 100_000_000,
-          eval_duration: 650_000_000,
-        },
-        {
-          model: 'llama3.2',
-          created_at: new Date(),
-          message: { role: 'assistant', content: ' three' },
-          done: true,
-          done_reason: 'stop',
-          eval_count: 15,
-          prompt_eval_count: 8,
-          total_duration: 1_000_000_000,
-          load_duration: 100_000_000,
-          prompt_eval_duration: 200_000_000,
-          eval_duration: 700_000_000,
-        },
-      ];
-
-      mockChatStream(mockStreamData);
-
-      const options: LanguageModelV4CallOptions = {
-        prompt: [{ role: 'user', content: [{ type: 'text', text: 'Test' }] }],
-      };
-
-      const { stream } = await model.doStream(options);
-      const chunks: LanguageModelV4StreamPart[] = await Array.fromAsync(stream);
-
-      // Extract text-related chunks
-      const textStart = chunks.find((chunk) => chunk.type === 'text-start') as {
-        id: string;
-      };
-      const textDeltas = chunks.filter(
-        (chunk) => chunk.type === 'text-delta',
-      ) as { id: string; delta: string }[];
-      const textEnd = chunks.find((chunk) => chunk.type === 'text-end') as {
-        id: string;
-      };
-
-      // All text parts should have the same ID
-      const expectedId = textStart.id;
-      expect(textStart.id).toBe(expectedId);
-      for (const delta of textDeltas) {
-        expect(delta.id).toBe(expectedId);
-      }
-      expect(textEnd.id).toBe(expectedId);
-
-      // Verify the deltas combine to form the complete text
-      const completeText = textDeltas.map((delta) => delta.delta).join('');
-      expect(completeText).toBe('Part two three');
     });
   });
 });
