+++ packages/ai-sdk-ollama/src/utils/object-generation-reliability.test.ts
@@ -4,7 +4,7 @@ import {
   cascadeRepairText,
   getRepairFunction,
   parseJSONWithRepair,
-} from './object-generation-reliability';
+} from './json-text-repair';
 
 describe('Enhanced JSON Repair', () => {
   describe('enhancedRepairText', () => {
